+++ pallets/loan/src/lib.rs
@@ -15,7 +15,7 @@ use frame_support::transactional;
 use frame_support::{ensure, Parameter};
 use frame_system::pallet_prelude::OriginFor;
 use frame_system::RawOrigin;
-use loan_type::{Asset, LoanType};
+use loan_type::LoanType;
 pub use pallet::*;
 #[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
@@ -25,6 +25,7 @@ use sp_runtime::{DispatchError, FixedPointNumber};
 use sp_std::convert::TryInto;
 #[cfg(feature = "std")]
 use std::fmt::Debug;
+use types::*;
 
 #[cfg(test)]
 mod mock;
@@ -32,148 +33,10 @@ mod mock;
 #[cfg(test)]
 mod tests;
 
+pub mod functions;
 mod loan_type;
 pub mod math;
-
-/// The data structure for storing pool nav details
-#[derive(Encode, Decode, Copy, Clone, PartialEq, Default)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
-pub struct NAVDetails<Amount> {
-	// this is the latest nav for the given pool.
-	// this will be updated on these scenarios
-	// 1. When we are calculating pool nav
-	// 2. when there is borrow or repay or write off on a loan under this pool
-	// So NAV could be
-	//	approximate when current time != last_updated
-	//	exact when current time == last_updated
-	latest_nav: Amount,
-
-	// this is the last time when the nav was calculated for the entire pool
-	last_updated: u64,
-}
-
-/// The data structure for storing a specific write off group
-#[derive(Encode, Decode, Copy, Clone, PartialEq, Default)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
-pub struct WriteOffGroup<Rate> {
-	/// percentage of outstanding debt we are going to write off on a loan
-	percentage: Rate,
-
-	/// number in days after the maturity has passed at which this write off group is valid
-	overdue_days: u64,
-}
-
-/// The data structure for storing loan info
-#[derive(Encode, Decode, Copy, Clone, PartialEq)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
-pub enum LoanStatus {
-	// this when asset is locked and loan nft is issued.
-	Issued,
-	// this is when loan is in active state. Either underwriters or oracles can move loan to this state
-	// by providing information like discount rates etc.. to loan
-	Active,
-	// loan is closed and asset nft is transferred back to borrower and loan nft is transferred back to loan module
-	Closed,
-}
-
-/// The data structure for storing loan info
-#[derive(Encode, Decode, Copy, Clone)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize, Debug))]
-pub struct LoanData<Rate, Amount, Asset> {
-	ceiling: Amount,
-	borrowed_amount: Amount,
-	rate_per_sec: Rate,
-	// accumulated rate till last_updated. more about this here - https://docs.makerdao.com/smart-contract-modules/rates-module
-	accumulated_rate: Rate,
-	// principal debt used to calculate the current outstanding debt.
-	// principal debt will change on every borrow and repay.
-	// Called principal debt instead of pie or normalized debt as mentioned here - https://docs.makerdao.com/smart-contract-modules/rates-module
-	// since its easier to look at it as principal amount borrowed and can be used to calculate final debt with the accumulated interest rate
-	principal_debt: Amount,
-	last_updated: u64,
-	asset: Asset,
-	status: LoanStatus,
-	loan_type: LoanType<Rate, Amount>,
-
-	// whether the loan written off by admin
-	// if so, we wont update the write off group on this loan further from permission less call
-	admin_written_off: bool,
-	// write off group index in the vec of write off groups
-	// none, the loan is not written off yet
-	// some(index), loan is written off and write off details are found under the given index
-	write_off_index: Option<u32>,
-}
-
-impl<Rate, Amount, Asset> LoanData<Rate, Amount, Asset>
-where
-	Rate: FixedPointNumber,
-	Amount: FixedPointNumber,
-{
-	/// returns the present value of the loan
-	/// note: this will use the accumulated_rate and last_updated from self
-	/// if you want the latest upto date present value, ensure these values are updated as well before calling this
-	fn present_value(&self) -> Option<Amount> {
-		// calculate current debt and present value
-		math::debt(self.principal_debt, self.accumulated_rate).and_then(|debt| {
-			self.loan_type
-				.present_value(debt, self.last_updated, self.rate_per_sec)
-		})
-	}
-
-	/// accrues rate and current debt from last updated until now
-	fn accrue(&self, now: u64) -> Option<(Rate, Amount)> {
-		// if the borrow amount is zero, then set accumulated rate to rate per sec so we start accumulating from now.
-		let maybe_rate = match self.borrowed_amount == Zero::zero() {
-			true => Some(self.rate_per_sec),
-			false => math::calculate_accumulated_rate::<Rate>(
-				self.rate_per_sec,
-				self.accumulated_rate,
-				now,
-				self.last_updated,
-			),
-		};
-
-		// calculate the current outstanding debt
-		let maybe_debt = maybe_rate
-			.and_then(|acc_rate| math::debt::<Amount, Rate>(self.principal_debt, acc_rate));
-
-		match (maybe_rate, maybe_debt) {
-			(Some(rate), Some(debt)) => Some((rate, debt)),
-			_ => None,
-		}
-	}
-
-	/// returns the present value of the loan adjusted to the write off group assigned to the loan if any
-	// pv = pv*(1 - write_off_percentage)
-	fn present_value_with_write_off(
-		&self,
-		write_off_groups: Vec<WriteOffGroup<Rate>>,
-	) -> Option<Amount> {
-		let maybe_present_value = self.present_value();
-		match self.write_off_index {
-			None => maybe_present_value,
-			Some(index) => maybe_present_value.and_then(|pv| {
-				write_off_groups
-					.get(index as usize)
-					// convert rate to amount
-					.and_then(|group| math::convert::<Rate, Amount>(group.percentage))
-					// calculate write off amount
-					.and_then(|write_off_percentage| pv.checked_mul(&write_off_percentage))
-					// calculate adjusted present value
-					.and_then(|write_off_amount| pv.checked_sub(&write_off_amount))
-			}),
-		}
-	}
-}
-
-/// type alias to Nonfungible ClassId type
-type ClassIdOf<T> =
-	<<T as Config>::NonFungible as Inspect<<T as frame_system::Config>::AccountId>>::ClassId;
-/// type alias to Nonfungible InstanceId type
-type InstanceIdOf<T> =
-	<<T as Config>::NonFungible as Inspect<<T as frame_system::Config>::AccountId>>::InstanceId;
-/// type alias to Nonfungible Asset
-pub type AssetOf<T> = Asset<<T as Config>::ClassId, <T as Config>::LoanId>;
+pub mod types;
 
 #[frame_support::pallet]
 pub mod pallet {
@@ -459,40 +322,7 @@ pub mod pallet {
 			loan_type: LoanType<T::Rate, T::Amount>,
 		) -> DispatchResult {
 			<T as Config>::AdminOrigin::ensure_origin(origin)?;
-
-			// check if the pool exists
-			pallet_pool::Pallet::<T>::check_pool(pool_id)?;
-
-			LoanInfo::<T>::try_mutate(pool_id, loan_id, |maybe_loan_info| -> DispatchResult {
-				let mut loan_info = maybe_loan_info.take().ok_or(Error::<T>::ErrMissingLoan)?;
-
-				// ensure loan is issued
-				ensure!(
-					loan_info.status == LoanStatus::Issued,
-					Error::<T>::ErrLoanIsActive
-				);
-
-				// ensure loan_type is valid
-				let now = Self::time_now()?;
-				ensure!(loan_type.is_valid(now), Error::<T>::ErrLoanValueInvalid);
-
-				// ensure rate_per_sec >= one
-				ensure!(rate_per_sec >= One::one(), Error::<T>::ErrLoanValueInvalid);
-
-				// calculate ceiling
-				let ceiling = loan_type.ceiling().ok_or(Error::<T>::ErrLoanTypeInvalid)?;
-				ensure!(ceiling > Zero::zero(), Error::<T>::ErrLoanValueInvalid);
-
-				// update the loan info
-				loan_info.rate_per_sec = rate_per_sec;
-				loan_info.ceiling = ceiling;
-				loan_info.status = LoanStatus::Active;
-				loan_info.loan_type = loan_type;
-				*maybe_loan_info = Some(loan_info);
-
-				Ok(())
-			})?;
-
+			Self::activate(pool_id, loan_id, rate_per_sec, loan_type)?;
 			Self::deposit_event(Event::<T>::LoanActivated(pool_id, loan_id));
 			Ok(())
 		}
@@ -514,29 +344,14 @@ pub mod pallet {
 		/// write off groups are always append only
 		#[pallet::weight(100_000)]
 		#[transactional]
-		pub fn add_write_off_group(
+		pub fn add_write_off_group_to_pool(
 			origin: OriginFor<T>,
 			pool_id: T::PoolId,
 			group: WriteOffGroup<T::Rate>,
 		) -> DispatchResult {
 			// ensure this is coming from an admin origin
 			<T as Config>::AdminOrigin::ensure_origin(origin)?;
-
-			// check if the pool exists
-			pallet_pool::Pallet::<T>::check_pool(pool_id)?;
-
-			// ensure write off percentage is not more than 100
-			ensure!(
-				group.percentage <= One::one(),
-				Error::<T>::ErrInvalidWriteOffGroup
-			);
-
-			// append new group
-			let index = PoolWriteOffGroups::<T>::mutate(pool_id, |write_off_groups| -> u32 {
-				write_off_groups.push(group);
-				// return the index of the write off group
-				(write_off_groups.len() - 1) as u32
-			});
+			let index = Self::add_write_off_group(pool_id, group)?;
 			Self::deposit_event(Event::<T>::WriteOffGroupAdded(pool_id, index));
 			Ok(())
 		}
@@ -577,499 +392,6 @@ pub mod pallet {
 			Self::deposit_event(Event::<T>::LoanWrittenOff(pool_id, loan_id, index));
 			Ok(())
 		}
-	}
-}
-
-impl<T: Config> Pallet<T> {
-	/// returns the account_id of the loan pallet
-	pub fn account_id() -> T::AccountId {
-		T::LoanPalletId::get().into_account()
-	}
-
-	// TODO(ved): make this a hook maybe
-	pub fn initialise_pool(
-		origin: OriginFor<T>,
-		pool_id: T::PoolId,
-		loan_nft_class_id: T::ClassId,
-	) -> DispatchResult {
-		// ensure admin is the origin
-		T::AdminOrigin::ensure_origin(origin)?;
-
-		// ensure pool is not initialised yet
-		ensure!(
-			!PoolToLoanNftClass::<T>::contains_key(pool_id),
-			Error::<T>::ErrPoolAlreadyInitialised
-		);
-
-		PoolToLoanNftClass::<T>::insert(pool_id, loan_nft_class_id);
-		LoanNftClassToPool::<T>::insert(loan_nft_class_id, pool_id);
-		Ok(())
-	}
-
-	/// check if the given loan belongs to the owner provided
-	fn check_loan_owner(
-		pool_id: T::PoolId,
-		loan_id: T::LoanId,
-		owner: T::AccountId,
-	) -> Result<AssetOf<T>, DispatchError> {
-		let loan_class_id =
-			PoolToLoanNftClass::<T>::get(pool_id).ok_or(Error::<T>::ErrPoolNotInitialised)?;
-		let got = T::NonFungible::owner(&loan_class_id.into(), &loan_id.into())
-			.ok_or(Error::<T>::ErrNFTOwnerNotFound)?;
-		ensure!(got == owner, Error::<T>::ErrNotAssetOwner);
-		Ok(Asset(loan_class_id, loan_id))
-	}
-
-	/// issues a new loan nft and returns the LoanID
-	fn issue(
-		pool_id: T::PoolId,
-		asset_owner: T::AccountId,
-		asset: AssetOf<T>,
-	) -> Result<T::LoanId, sp_runtime::DispatchError> {
-		// check if the nft belongs to owner
-		let (asset_class_id, instance_id) = asset.destruct();
-		let owner = T::NonFungible::owner(&asset_class_id.into(), &instance_id.into())
-			.ok_or(Error::<T>::ErrNFTOwnerNotFound)?;
-		ensure!(owner == asset_owner, Error::<T>::ErrNotAssetOwner);
-
-		// check if the registry is not an loan nft registry
-		ensure!(
-			!LoanNftClassToPool::<T>::contains_key(asset_class_id),
-			Error::<T>::ErrNotAValidAsset
-		);
-
-		// create new loan nft
-		let loan_pallet_account: T::AccountId = T::LoanPalletId::get().into_account();
-		let nonce = NextLoanId::<T>::get();
-		let loan_id: T::LoanId = nonce.into();
-		let loan_class_id =
-			PoolToLoanNftClass::<T>::get(pool_id).ok_or(Error::<T>::ErrPoolNotInitialised)?;
-		T::NonFungible::mint_into(&loan_class_id.into(), &loan_id.into(), &owner)?;
-
-		// lock asset nft
-		T::NonFungible::transfer(
-			&asset_class_id.into(),
-			&instance_id.into(),
-			&loan_pallet_account,
-		)?;
-		let timestamp = Self::time_now()?;
-
-		// update the next token nonce
-		let next_loan_id = nonce
-			.checked_add(1)
-			.ok_or(Error::<T>::ErrNftTokenNonceOverflowed)?;
-		NextLoanId::<T>::set(next_loan_id);
-
-		// create loan info
-		LoanInfo::<T>::insert(
-			pool_id,
-			loan_id,
-			LoanData {
-				ceiling: Zero::zero(),
-				borrowed_amount: Zero::zero(),
-				rate_per_sec: Zero::zero(),
-				accumulated_rate: One::one(),
-				principal_debt: Zero::zero(),
-				last_updated: timestamp,
-				status: LoanStatus::Issued,
-				loan_type: Default::default(),
-				admin_written_off: false,
-				write_off_index: None,
-				asset,
-			},
-		);
-		Ok(loan_id)
-	}
-
-	fn close(
-		pool_id: T::PoolId,
-		loan_id: T::LoanId,
-		owner: T::AccountId,
-	) -> Result<AssetOf<T>, DispatchError> {
-		// ensure owner is the loan nft owner
-		let loan_nft = Self::check_loan_owner(pool_id, loan_id, owner.clone())?;
-
-		LoanInfo::<T>::try_mutate(
-			pool_id,
-			loan_id,
-			|maybe_loan_info| -> Result<AssetOf<T>, DispatchError> {
-				let mut loan_info = maybe_loan_info.take().ok_or(Error::<T>::ErrMissingLoan)?;
-
-				// ensure loan is active
-				ensure!(
-					loan_info.status == LoanStatus::Active,
-					Error::<T>::ErrLoanNotActive
-				);
-
-				// ensure debt is all paid
-				// we just need to ensure principal debt is zero
-				// if not, we check if the loan is written of 100%
-				match (
-					loan_info.principal_debt == Zero::zero(),
-					loan_info.write_off_index,
-				) {
-					// debt is cleared
-					(true, _) => Ok(()),
-					// debt not cleared and loan not written off
-					(_, None) => Err(Error::<T>::ErrLoanNotRepaid),
-					// debt not cleared but loan is written off
-					// if written off completely, then we can close it
-					(_, Some(write_off_index)) => {
-						let groups = PoolWriteOffGroups::<T>::get(pool_id);
-						let group = groups
-							.get(write_off_index as usize)
-							.ok_or(Error::<T>::ErrInvalidWriteOffGroupIndex)?;
-						ensure!(group.percentage == One::one(), Error::<T>::ErrLoanNotRepaid);
-						Ok(())
-					}
-				}?;
-
-				// transfer asset to owner
-				let asset = loan_info.asset;
-				let (asset_class_id, instance_id) = asset.destruct();
-				T::NonFungible::transfer(&asset_class_id.into(), &instance_id.into(), &owner)?;
-
-				// transfer loan nft to loan pallet
-				// ideally we should burn this but we do not have a function to burn them yet.
-				// TODO(ved): burn loan nft when the functionality is available
-				let (loan_class_id, loan_id) = loan_nft.destruct();
-				T::NonFungible::transfer(
-					&loan_class_id.into(),
-					&loan_id.into(),
-					&Self::account_id(),
-				)?;
-
-				// update loan status
-				loan_info.status = LoanStatus::Closed;
-				*maybe_loan_info = Some(loan_info);
-				Ok(asset)
-			},
-		)
-	}
-
-	fn borrow_amount(
-		pool_id: T::PoolId,
-		loan_id: T::LoanId,
-		owner: T::AccountId,
-		amount: T::Amount,
-	) -> DispatchResult {
-		// ensure owner is the loan owner
-		Self::check_loan_owner(pool_id, loan_id, owner.clone())?;
-
-		LoanInfo::<T>::try_mutate(pool_id, loan_id, |maybe_loan_info| -> DispatchResult {
-			let mut loan_info = maybe_loan_info.take().ok_or(Error::<T>::ErrMissingLoan)?;
-
-			// ensure loan is active
-			ensure!(
-				loan_info.status == LoanStatus::Active,
-				Error::<T>::ErrLoanNotActive
-			);
-
-			// ensure maturity date has not passed if the loan has a maturity date
-			let now: u64 = Self::time_now()?;
-			let valid = match loan_info.loan_type.maturity_date() {
-				// loan has a maturity date
-				Some(md) => md > now,
-				// no maturity date, so continue as is
-				None => true,
-			};
-			ensure!(valid, Error::<T>::ErrLoanMaturityDatePassed);
-
-			// ensure borrow amount is positive
-			ensure!(amount.is_positive(), Error::<T>::ErrLoanValueInvalid);
-
-			// check for ceiling threshold
-			ensure!(
-				amount + loan_info.borrowed_amount <= loan_info.ceiling,
-				Error::<T>::ErrLoanCeilingReached
-			);
-
-			// get previous present value so that we can update the nav accordingly
-			let old_pv = loan_info
-				.present_value()
-				.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-
-			// calculate accumulated rate and outstanding debt
-			let (accumulated_rate, debt) = loan_info
-				.accrue(now)
-				.ok_or(Error::<T>::ErrLoanAccrueFailed)?;
-
-			let new_borrowed_amount = loan_info
-				.borrowed_amount
-				.checked_add(&amount)
-				.ok_or(Error::<T>::ErrAddAmountOverflow)?;
-
-			// calculate new principal debt with adjustment amount
-			let principal_debt = math::calculate_principal_debt::<T::Amount, T::Rate>(
-				debt,
-				math::Adjustment::Inc(amount),
-				accumulated_rate,
-			)
-			.ok_or(Error::<T>::ErrPrincipalDebtOverflow)?;
-
-			// update loan
-			loan_info.borrowed_amount = new_borrowed_amount;
-			loan_info.last_updated = now;
-			loan_info.accumulated_rate = accumulated_rate;
-			loan_info.principal_debt = principal_debt;
-			let new_pv = loan_info
-				.present_value()
-				.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-			Self::update_nav_with_updated_present_value(pool_id, new_pv, old_pv)?;
-			pallet_pool::Pallet::<T>::borrow_currency(
-				pool_id,
-				RawOrigin::Signed(Self::account_id()).into(),
-				owner,
-				amount.into(),
-			)?;
-			*maybe_loan_info = Some(loan_info);
-			Ok(())
-		})
-	}
-
-	fn update_nav_with_updated_present_value(
-		pool_id: T::PoolId,
-		new_pv: T::Amount,
-		old_pv: T::Amount,
-	) -> Result<(), DispatchError> {
-		// calculate new diff from the old and new present value and update the nav accordingly
-		PoolNAV::<T>::try_mutate(pool_id, |maybe_nav_details| -> Result<(), DispatchError> {
-			let mut nav = maybe_nav_details.take().unwrap_or_default();
-			let new_nav = match new_pv > old_pv {
-				// borrow
-				true => new_pv
-					.checked_sub(&old_pv)
-					.and_then(|positive_diff| nav.latest_nav.checked_add(&positive_diff)),
-				// repay since new pv is less than old
-				false => old_pv
-					.checked_sub(&new_pv)
-					.and_then(|negative_diff| nav.latest_nav.checked_sub(&negative_diff)),
-			}
-			.ok_or(Error::<T>::ErrAddAmountOverflow)?;
-			nav.latest_nav = new_nav;
-			*maybe_nav_details = Some(nav);
-			Ok(())
-		})
-	}
-
-	fn repay_amount(
-		pool_id: T::PoolId,
-		loan_id: T::LoanId,
-		owner: T::AccountId,
-		amount: T::Amount,
-	) -> Result<T::Amount, DispatchError> {
-		// ensure owner is the loan owner
-		Self::check_loan_owner(pool_id, loan_id, owner.clone())?;
-
-		LoanInfo::<T>::try_mutate(
-			pool_id,
-			loan_id,
-			|maybe_loan_info| -> Result<T::Amount, DispatchError> {
-				let mut loan_info = maybe_loan_info.take().ok_or(Error::<T>::ErrMissingLoan)?;
-
-				// ensure loan is active
-				ensure!(
-					loan_info.status == LoanStatus::Active,
-					Error::<T>::ErrLoanNotActive
-				);
-
-				// ensure repay amount is positive
-				ensure!(amount.is_positive(), Error::<T>::ErrLoanValueInvalid);
-
-				// calculate old present_value
-				let old_pv = loan_info
-					.present_value()
-					.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-
-				// calculate new accumulated rate
-				let now: u64 = Self::time_now()?;
-				let (accumulated_rate, debt) = loan_info
-					.accrue(now)
-					.ok_or(Error::<T>::ErrLoanAccrueFailed)?;
-
-				// ensure amount is not more than current debt
-				let mut repay_amount = amount;
-				if repay_amount > debt {
-					repay_amount = debt
-				}
-
-				// calculate new principal debt with repaid amount
-				let principal_debt = math::calculate_principal_debt::<T::Amount, T::Rate>(
-					debt,
-					math::Adjustment::Dec(repay_amount),
-					accumulated_rate,
-				)
-				.ok_or(Error::<T>::ErrAddAmountOverflow)?;
-
-				loan_info.last_updated = now;
-				loan_info.accumulated_rate = accumulated_rate;
-				loan_info.principal_debt = principal_debt;
-				let new_pv = loan_info
-					.present_value()
-					.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-				Self::update_nav_with_updated_present_value(pool_id, new_pv, old_pv)?;
-				pallet_pool::Pallet::<T>::repay_currency(
-					pool_id,
-					RawOrigin::Signed(Self::account_id()).into(),
-					owner,
-					repay_amount.into(),
-				)?;
-				*maybe_loan_info = Some(loan_info);
-				Ok(repay_amount)
-			},
-		)
-	}
-
-	fn time_now() -> Result<u64, DispatchError> {
-		let nowt = T::Time::now();
-		TryInto::<u64>::try_into(nowt).map_err(|_| Error::<T>::ErrEpochTimeOverflow.into())
-	}
-
-	/// accrues rate and debt of a given loan and updates it
-	/// returns the present value of the loan accounting any write offs
-	fn accrue_and_update_loan(
-		pool_id: T::PoolId,
-		loan_id: T::LoanId,
-		now: u64,
-		write_off_groups: Vec<WriteOffGroup<T::Rate>>,
-	) -> Result<T::Amount, DispatchError> {
-		LoanInfo::<T>::try_mutate(
-			pool_id,
-			loan_id,
-			|maybe_loan_data| -> Result<T::Amount, DispatchError> {
-				let mut loan_data = maybe_loan_data.take().ok_or(Error::<T>::ErrMissingLoan)?;
-				// if the loan is not active, then skip updating and return PV as zero
-				if loan_data.status != LoanStatus::Active {
-					*maybe_loan_data = Some(loan_data);
-					return Ok(Zero::zero());
-				}
-
-				let (acc_rate, _debt) = loan_data
-					.accrue(now)
-					.ok_or(Error::<T>::ErrLoanAccrueFailed)?;
-				loan_data.last_updated = now;
-				loan_data.accumulated_rate = acc_rate;
-				let present_value = loan_data
-					.present_value_with_write_off(write_off_groups)
-					.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-				*maybe_loan_data = Some(loan_data);
-				Ok(present_value)
-			},
-		)
-	}
-
-	/// updates nav for the given pool and returns the latest NAV at this instant
-	fn update_nav_of_pool(pool_id: T::PoolId) -> Result<T::Amount, DispatchError> {
-		let now = Self::time_now()?;
-		let write_off_groups = PoolWriteOffGroups::<T>::get(pool_id);
-		let nav = LoanInfo::<T>::iter_key_prefix(pool_id).try_fold(
-			Zero::zero(),
-			|sum, loan_id| -> Result<T::Amount, DispatchError> {
-				let pv =
-					Self::accrue_and_update_loan(pool_id, loan_id, now, write_off_groups.clone())?;
-				sum.checked_add(&pv)
-					.ok_or(Error::<T>::ErrLoanAccrueFailed.into())
-			},
-		)?;
-		PoolNAV::<T>::insert(
-			pool_id,
-			NAVDetails {
-				latest_nav: nav,
-				last_updated: now,
-			},
-		);
-		Ok(nav)
-	}
-
-	/// writes off a given unhealthy loan
-	/// if override_write_off_index is Some, this is a admin action and loan override flag is set
-	/// if loan is already overridden and override_write_off_index is None, we return error
-	/// if loan is still healthy, we return an error
-	/// loan is accrued and nav is updated accordingly
-	/// returns new write off index applied to loan
-	fn write_off(
-		pool_id: T::PoolId,
-		loan_id: T::LoanId,
-		override_write_off_index: Option<u32>,
-	) -> Result<u32, DispatchError> {
-		LoanInfo::<T>::try_mutate(
-			pool_id,
-			loan_id,
-			|maybe_loan_data| -> Result<u32, DispatchError> {
-				let mut loan_data = maybe_loan_data.take().ok_or(Error::<T>::ErrMissingLoan)?;
-				// ensure loan is active
-				ensure!(
-					loan_data.status == LoanStatus::Active,
-					Error::<T>::ErrLoanNotActive
-				);
-
-				let now = Self::time_now()?;
-
-				// ensure loan was not overwritten by admin and try to fetch a valid write off group for loan
-				let write_off_groups = PoolWriteOffGroups::<T>::get(pool_id);
-				let write_off_group_index =
-					match (loan_data.admin_written_off, override_write_off_index) {
-						// admin is trying to write off
-						(_admin_written_off, Some(index)) => {
-							// check if the write off group exists
-							write_off_groups
-								.get(index as usize)
-								.ok_or(Error::<T>::ErrInvalidWriteOffGroupIndex)?;
-							loan_data.admin_written_off = true;
-							Ok(index)
-						}
-						(admin_written_off, None) => {
-							// non-admin is trying to write off but admin already did. So error out
-							if admin_written_off {
-								return Err(Error::<T>::ErrLoanWrittenOffByAdmin.into());
-							}
-
-							let maturity_date = loan_data
-								.loan_type
-								.maturity_date()
-								.ok_or(Error::<T>::ErrLoanTypeInvalid)?;
-
-							// ensure loan's maturity date has passed
-							ensure!(now > maturity_date, Error::<T>::ErrLoanHealthy);
-
-							// not written off by admin, and non admin trying to write off, then
-							// fetch the best write group available for this loan
-							math::valid_write_off_group(
-								maturity_date,
-								now,
-								write_off_groups.clone(),
-							)
-							.ok_or(Error::<T>::ErrNoValidWriteOffGroup)
-						}
-					}?;
-
-				// get old present value accounting for any write offs
-				let old_pv = loan_data
-					.present_value_with_write_off(write_off_groups.clone())
-					.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-
-				// accrue and calculate the new present value with current chosen write off
-				let (accumulated_rate, _current_debt) = loan_data
-					.accrue(now)
-					.ok_or(Error::<T>::ErrLoanAccrueFailed)?;
-
-				loan_data.accumulated_rate = accumulated_rate;
-				loan_data.last_updated = now;
-				loan_data.write_off_index = Some(write_off_group_index);
-
-				// calculate updated write off adjusted present value
-				let new_pv = loan_data
-					.present_value_with_write_off(write_off_groups)
-					.ok_or(Error::<T>::ErrLoanPresentValueFailed)?;
-
-				// update nav
-				Self::update_nav_with_updated_present_value(pool_id, new_pv, old_pv)?;
-
-				// update loan data
-				*maybe_loan_data = Some(loan_data);
-				Ok(write_off_group_index)
-			},
-		)
 	}
 }
